refactor modal screen to allow for add connection modal
