+++ packages/frontend/src/components/Editor/index.tsx
@@ -36,7 +36,7 @@ import {
 } from '@/helpers/toolbox'
 import useAuthentication from '@/hooks/useAuthentication'
 
-import EmptyFlowStepHeaderModal from '../EmptyFlowStepHeader/EmptyFlowStepHeaderModal'
+import FlowStepConfigurationModal from '../FlowStepConfigurationModal'
 
 // TODO(mal): Remove this comment before merging this main PR
 interface AddStepButtonProps {
@@ -100,7 +100,7 @@ function AddStepButton(props: AddStepButtonProps): JSX.Element {
         </AbsoluteCenter>
       </Box>
 
-      <EmptyFlowStepHeaderModal
+      <FlowStepConfigurationModal
         isOpen={isOpen}
         onClose={onClose}
         isTrigger={false} // Can only add an action all the time
